+++ BT2_android/MyApplication/app/src/main/java/com/example/myapplication/Adapter/CityAdapter.java
@@ -1,5 +1,5 @@
 // CityAdapter.java
-package com.example.myapplication;
+package com.example.myapplication.Adapter;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -7,6 +7,10 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.myapplication.Model.City;
+import com.example.myapplication.R;
+
 import java.util.List;
 
 public class CityAdapter extends RecyclerView.Adapter<CityAdapter.CityViewHolder> {
