+++ BT2_android/MyApplication/app/src/main/java/com/example/myapplication/Model/City.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.Model;
 
 public class City {
     private String name;
